requirements.txt
flake8
GitPython
Pillow
python-dotenv
requests
PyGitHub
